mongodb/requirements.txt
pymongo==4.6.1
python-dotenv==1.0.0
openai>=1.10.0
langchain==0.1.12
langchain_openai==0.0.8
langchain_community>=0.0.28
langchain_mongodb>=0.0.1
chromadb==0.4.18
tiktoken==0.5.2
python-magic==0.4.27
beautifulsoup4==4.12.2
requests==2.31.0
unstructured>=0.10.30
pdf2image>=1.16.3
pdfminer.six>=20221105
python-docx>=0.8.11
python-pptx>=0.6.21
